move everything to Content/Swamp
