+++ Assets/Swamp/NPCs/GiantEel.png
