+++ Assets/Swamp/NPCs/GiantEelBody.png
