+++ Assets/Swamp/NPCs/GiantEelTail.png
