+++ Assets/Swamp/NPCs/Gores/DragonFly_Body.png
