+++ Assets/Swamp/NPCs/Gores/DragonFly_Head.png
