+++ Assets/Swamp/NPCs/Gores/DragonFly_Mandible.png
