+++ Assets/Swamp/NPCs/Gores/DragonFly_TailTip.png
