+++ Assets/Swamp/NPCs/Gores/DragonFly_WingSmall.png
