+++ Assets/Swamp/NPCs/Gores/DragonFly_Wings.png
