+++ Assets/Swamp/NPCs/Gores/SwampCroc_0.png
