+++ Assets/Swamp/NPCs/Gores/SwampCroc_1.png
