+++ Assets/Swamp/NPCs/Gores/SwampCroc_2.png
