+++ Assets/Swamp/NPCs/Gores/SwampCroc_3.png
