+++ Assets/Swamp/NPCs/Gores/SwampCroc_4.png
