+++ Assets/Swamp/NPCs/Mudsquit.png
