+++ Assets/Swamp/NPCs/SwampCroc.png
