+++ Content/Swamp/NPCs/Mudsquit.cs
@@ -12,7 +12,7 @@ using Terraria.GameContent;
 using Terraria.GameContent.Bestiary;
 using Terraria.ID;
 
-namespace PathOfTerraria.Content.NPCs.Mapping.Swamp;
+namespace PathOfTerraria.Content.Swamp.NPCs;
 
 internal class Mudsquit : ModNPC
 {
