+++ Content/Swamp/SwampArea.cs
@@ -1,11 +1,12 @@
 ﻿using PathOfTerraria.Common.Encounters;
-using PathOfTerraria.Common.Subworlds.MappingAreas.SwampAreaContent;
+using PathOfTerraria.Common.Subworlds;
+using PathOfTerraria.Common.Subworlds.MappingAreas;
 using PathOfTerraria.Common.Tiles.FramingKinds;
 using PathOfTerraria.Common.World.Generation;
 using PathOfTerraria.Common.World.Passes;
 using PathOfTerraria.Common.World.Utilities;
 using PathOfTerraria.Content.NPCs.Mapping.Desert.SunDevourer;
-using PathOfTerraria.Content.NPCs.Mapping.Swamp;
+using PathOfTerraria.Content.Swamp.NPCs;
 using PathOfTerraria.Content.Tiles.Maps.Swamp;
 using PathOfTerraria.Content.Walls;
 using PathOfTerraria.Utilities;
@@ -20,7 +21,7 @@ using Terraria.Localization;
 using Terraria.Utilities;
 using Terraria.WorldBuilding;
 
-namespace PathOfTerraria.Common.Subworlds.MappingAreas;
+namespace PathOfTerraria.Content.Swamp;
 
 #nullable enable
 
@@ -314,7 +315,7 @@ internal class SwampArea : MappingWorld, IExplorationWorld, IOverrideBiome
 		for (int i = 0; i < 150; ++i)
 		{
 			Point16 firstPoint = Random.Next(leaves);
-			Vector2 one = firstPoint.ToVector2() + OpenExtensions.GetDirectionRandom(OpenExtensions.GetUnsolidOpenings(firstPoint.X, firstPoint.Y, false, false)).ToVector2();
+			Vector2 one = firstPoint.ToVector2() + OpenExtensions.GetUnsolidOpenings(firstPoint.X, firstPoint.Y, false, false).GetDirectionRandom().ToVector2();
 
 			Point16 secondPoint;
 
@@ -323,7 +324,7 @@ internal class SwampArea : MappingWorld, IExplorationWorld, IOverrideBiome
 				secondPoint = Random.Next(leaves);
 			} while (firstPoint == secondPoint || one.DistanceSQ(secondPoint.ToVector2()) > 700 * 700);
 
-			Vector2 two = secondPoint.ToVector2() + OpenExtensions.GetDirectionRandom(OpenExtensions.GetUnsolidOpenings(secondPoint.X, secondPoint.Y, false, false)).ToVector2();
+			Vector2 two = secondPoint.ToVector2() + OpenExtensions.GetUnsolidOpenings(secondPoint.X, secondPoint.Y, false, false).GetDirectionRandom().ToVector2();
 			float distance = two.Distance(one);
 			Vector2[] vine = Tunnel.GenerateBezier([one, Vector2.Lerp(one, two, 0.5f) + new Vector2(0, Random.NextFloat(distance * 0.33f, distance * 0.5f)), two], 0.8f, 0);
 			float baseStrength = Random.NextFloat(0.8f, 1.5f) * (distance / 500f);
@@ -352,7 +353,7 @@ internal class SwampArea : MappingWorld, IExplorationWorld, IOverrideBiome
 
 					for (int v = 0; v < length; ++v)
 					{
-						GenPlacement.WallCircle(pos + new Vector2(0, v), 1.5f - (v / length), WallID.LivingLeaf);
+						GenPlacement.WallCircle(pos + new Vector2(0, v), 1.5f - v / length, WallID.LivingLeaf);
 
 						if (Random.NextBool(8))
 						{
@@ -477,7 +478,7 @@ internal class SwampArea : MappingWorld, IExplorationWorld, IOverrideBiome
 			Point16 boulders = decor[1].Dequeue();
 			bool underwater = Main.tile[boulders].LiquidAmount > 0;
 			int chance = underwater ? 4 : 2;
-			int id = !Random.NextBool() ? (underwater ? ModContent.TileType<DeepMoss>() : ModContent.TileType<SwampMoss>()) : TileID.Mudstone;
+			int id = !Random.NextBool() ? underwater ? ModContent.TileType<DeepMoss>() : ModContent.TileType<SwampMoss>() : TileID.Mudstone;
 			GenPlacement.GenOval(boulders.ToVector2(), Random.NextFloat(2, 8), Random.NextFloat(MathHelper.PiOver2 - 0.3f, MathHelper.PiOver2 + 0.3f) * (Random.NextBool() ? -1 : 1),
 				id, (x, y) => noise.GetNoise(x, y) * 8);
 		}
@@ -542,7 +543,7 @@ internal class SwampArea : MappingWorld, IExplorationWorld, IOverrideBiome
 		{
 			x++;
 
-			if (Random.NextBool(150) || (x - lastX) > 400)
+			if (Random.NextBool(150) || x - lastX > 400)
 			{
 				if (x < Main.maxTilesX - 400)
 				{
